% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnect.R
\name{pool}
\alias{pool}
\title{Cria uma conexão de pool com um banco de dados PostgreSQL}
\usage{
pool()
}
\value{
Um objeto de conexão de pool.
}
\description{
Esta função cria e retorna uma conexão de pool com um banco de dados PostgreSQL.
O pool de conexão ajuda a gerenciar eficientemente as conexões de banco de dados.
}
\examples{
pool()
# Retorna: Uma conexão de pool com o banco de dados PostgreSQL.
}
